Fix Concorde timeout, NumPy indexing error, and solver worker hangs
Three robustness issues discovered while running ViTSP on macOS ARM
with open-source VLMs via Fireworks AI (qwen3p7-plus):

1. Subprocess-based Concorde timeout (exact_concorde.py)
   Concorde's time_bound parameter is unreliable (pyconcorde#17).
   Refactored Concorde.optimize() to run solver.solve() in a child
   process via mp.Process with a hard timeout. The parent terminates
   the subprocess if it exceeds timelimit + 5 seconds, preventing
   indefinite hangs during branch-and-bound.

2. Integer dtype for NumPy index arrays (tsp.py)
   generate_new_distance_mat_v3() and its CuPy variant crash with
   "IndexError: arrays used as indices must be of integer type" when
   segment/node representative arrays contain non-integer values.
   Fixed by explicitly casting to np.intp / cp.intp.

3. Timeout and error handling in solver workers (solver.py)
   - process_subTSP: wrapped reformulate_and_solve_subTSP in
     try/except so failed subproblems return the current route
     instead of crashing the child process silently.
   - Three busy-wait result-collection loops in solver_master,
     subproblem_solver, and subproblem_verifier now have a hard
     deadline (SolverTimeLimit + 20s). Stuck child processes are
     terminated when the deadline is reached.

Also adds .gitignore and removes tracked __pycache__/ files.

+++ src/LLM_TSP/tsp.py
@@ -270,10 +270,10 @@ class SubTSP():
 
         self.distance_mat = np.zeros((total, total), dtype=self.global_distance_mat.dtype)
 
-        # === Precompute representative nodes ===
-        seg_out_reps = np.array([self.route_segments[self.segment_map[i]][-1] for i in range(num_segments)])
-        seg_in_reps = np.array([self.route_segments[self.segment_map[i]][0] for i in range(num_segments)])
-        node_reps = np.array([self.node_map[i] for i in range(num_segments, total)])
+        # === Precompute representative nodes (ensure integer dtype for indexing) ===
+        seg_out_reps = np.array([self.route_segments[self.segment_map[i]][-1] for i in range(num_segments)], dtype=np.intp)
+        seg_in_reps = np.array([self.route_segments[self.segment_map[i]][0] for i in range(num_segments)], dtype=np.intp)
+        node_reps = np.array([self.node_map[i] for i in range(num_segments, total)], dtype=np.intp)
 
         # === Fill segment->segment block (top-left) ===
         self.distance_mat[:num_segments, :num_segments] = self.global_distance_mat[np.ix_(seg_out_reps, seg_in_reps)]
@@ -364,10 +364,10 @@ class SubTSP():
         global_mat = cp.asarray(self.global_distance_mat) if not isinstance(self.global_distance_mat,
                                                                             cp.ndarray) else self.global_distance_mat
 
-        # Precompute representative node indices
-        seg_out_reps = cp.array([self.route_segments[self.segment_map[i]][-1] for i in range(num_segments)])
-        seg_in_reps = cp.array([self.route_segments[self.segment_map[i]][0] for i in range(num_segments)])
-        node_reps = cp.array([self.node_map[i] for i in range(num_segments, total)])
+        # Precompute representative node indices (ensure integer dtype for indexing)
+        seg_out_reps = cp.array([self.route_segments[self.segment_map[i]][-1] for i in range(num_segments)], dtype=cp.intp)
+        seg_in_reps = cp.array([self.route_segments[self.segment_map[i]][0] for i in range(num_segments)], dtype=cp.intp)
+        node_reps = cp.array([self.node_map[i] for i in range(num_segments, total)], dtype=cp.intp)
 
         # Fill blocks
         self.distance_mat[:num_segments, :num_segments] = global_mat[cp.ix_(seg_out_reps, seg_in_reps)]
